Revise introduction
There are two parts that are not linked with each other. You need to have a narrative. This is why we are doing this, this is what we are doing, this is how we do it, this is what we need to prove (that this works well or is better), and so on. It's much better than before, but you need to write a bit more that can be really read as above. Refs #4